services/python-services/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.129.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
lxml==6.0.2
openai==2.21.0
orjson==3.11.7
packaging==24.2
pillow==12.1.1
pinecone==8.0.1
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.22
python-pptx==1.0.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
xlsxwriter==3.2.9
fastapi>=0.115
uvicorn>=0.34
pymupdf>=1.25
python-multipart>=0.0.18
python-docx>=0.8.11
python-pptx>=0.6.21
google-genai>=0.2.0
python-dotenv>=1.0.0
